# api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP API of the pylovo workflow (database, regions, data, generation, results) for the GridPlanner UI.",
    "title": "pylovo-api",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/grids/{grid_result_id}": {
      "get": {
        "description": "One grid with KPIs, feeders and GeoJSON layers.",
        "operationId": "grid_api_grids__grid_result_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "grid_result_id",
            "required": true,
            "schema": {
              "title": "Grid Result Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offsets",
            "required": false,
            "schema": {
              "default": true,
              "title": "Offsets",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Grid Api Grids  Grid Result Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grid",
        "tags": [
          "results"
        ]
      }
    },
...
    "/api/grids/{grid_result_id}/lod2": {
      "get": {
        "description": "LoD2 models of the grid's buildings from ``citydb`` as a binary mesh (format: :func:`pylovo_api.lod2.encode`).\n\nBuildings without LoD2, or a database without ``citydb``, give an empty mesh and never an error:\nthe map then shows those buildings in 2D.",
        "operationId": "grid_lod2_api_grids__grid_result_id__lod2_get",
        "parameters": [
          {
            "in": "path",
            "name": "grid_result_id",
            "required": true,
            "schema": {
              "title": "Grid Result Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grid Lod2",
        "tags": [
          "results"
        ]
      }
    },
    "/api/grids/{grid_result_id}/pandapower.json": {
      "get": {
        "description": "The stored pandapower network (load it with ``pandapower.from_json``).",
        "operationId": "grid_download_api_grids__grid_result_id__pandapower_json_get",
        "parameters": [
          {
            "in": "path",
            "name": "grid_result_id",
            "required": true,
            "schema": {
              "title": "Grid Result Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grid Download",
        "tags": [
          "results"
        ]
      }
    },
    "/api/grids/{grid_result_id}/powerflow": {
      "post": {
        "description": "Run a pandapower power flow now (nothing is stored).",
        "operationId": "grid_powerflow_api_grids__grid_result_id__powerflow_post",
        "parameters": [
          {
            "in": "path",
            "name": "grid_result_id",
            "required": true,
            "schema": {
              "title": "Grid Result Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "load_scaling",
            "required": false,
            "schema": {
              "default": 1.0,
              "exclusiveMinimum": 0,
              "maximum": 5,
              "title": "Load Scaling",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Grid Powerflow Api Grids  Grid Result Id  Powerflow Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grid Powerflow",
        "tags": [
          "results"
        ]
      }
    },
...
    "/api/results/{version_id}/summary": {
      "get": {
        "description": "KPIs and chart data of a version (all PLZ, or the given ones).",
        "operationId": "summary_api_results__version_id__summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "plz",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plz"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Summary Api Results  Version Id  Summary Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Summary",
        "tags": [
          "results"
        ]
      }
    },
    "/api/results/{version_id}/{plz}/map": {
      "get": {
        "description": "All grids of one PLZ as GeoJSON layers (cables, stations, buildings).",
        "operationId": "overview_api_results__version_id___plz__map_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "plz",
            "required": true,
            "schema": {
              "title": "Plz",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offsets",
            "required": false,
            "schema": {
              "default": true,
              "title": "Offsets",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Overview Api Results  Version Id   Plz  Map Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Overview",
        "tags": [
          "results"
        ]
      }
    },
...
    "/api/versions/compare": {
      "get": {
        "description": "Generation parameters of two versions side by side.",
        "operationId": "compare_api_versions_compare_get",
        "parameters": [
          {
            "in": "query",
            "name": "a",
            "required": true,
            "schema": {
              "title": "A",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "b",
            "required": true,
            "schema": {
              "title": "B",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Compare Api Versions Compare Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare",
        "tags": [
          "results"
        ]
      }
    },
    "/api/versions/{version_id}": {
      "get": {
        "description": "One version including its stored generation parameters.",
        "operationId": "version_api_versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Version Api Versions  Version Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Version",
        "tags": [
          "results"
        ]
      }
    },
...
  }
}